pyproject.toml
# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["core/**","authentication/**","coach/**","products/**","events/**","quizzes/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TC",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "UP038", # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
  "UP038",
]

[tool.ruff.lint.isort]
force-single-line = true

[project]
name = "coach-trusted-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "python-slugify (==8.0.4)",
    "pillow (==11.3.0)",
    "argon2-cffi (==25.1.0)",
    "whitenoise (==6.9.0)",
    "redis (==6.2.0)",
    "hiredis (==3.2.1)",
    "celery (==5.5.3)",
    "django-celery-beat (==2.8.1)",
    "flower (==2.0.1)",
    "uvicorn[standard] (==0.35.0)",
    "uvicorn-worker (==0.3.0)",
    "django (==4.2.20)",
    "django-environ (==0.12.0)",
    "django-model-utils (==5.0.0)",
    "django-allauth[mfa] (==65.10.0)",
    "django-crispy-forms (==2.4)",
    "crispy-bootstrap5 (==2025.6)",
    "django-redis (==5.4.0)",
    "django-extensions (==4.1)",
    "django-unfold (==0.63.0)",
    "djangorestframework (==3.16.0)",
    "django-cors-headers (==4.7.0)",
    "djangorestframework-simplejwt (==5.5.1)",
    "drf-spectacular (==0.28.0)",
    "django-filter (==25.1)",
    "firebase-admin (==6.9.0)",
    "pytest-cov (==6.2.1)",
    "django-ckeditor-5 (==0.2.18)",
    "werkzeug[watchdog] (==3.1.3)",
    "ipdb (==0.13.13)",
    "psycopg[c] (==3.2.9)",
    "watchfiles (==1.1.0)",
    "mypy (==1.17.0)",
    "django-stubs[compatible-mypy] (==5.2.2)",
    "pytest (==8.4.1)",
    "pytest-sugar (==1.0.0)",
    "djangorestframework-stubs (==3.16.1)",
    "sphinx (==8.3.0)",
    "sphinx-autobuild (==2024.10.3)",
    "ruff (==0.12.4)",
    "coverage (==7.9.2)",
    "djlint (==1.36.4)",
    "pre-commit (==4.2.0)",
    "factory-boy (==3.3.3)",
    "django-debug-toolbar (==5.2.0)",
    "django-coverage-plugin (==3.1.1)",
    "pytest-django (==4.11.1)",
    "gunicorn (==23.0.0)",
    "sentry-sdk (==2.33.2)",
    "django-anymail (==13.0)",
    "django-storages (==1.14.6)",
    "boto3 (==1.39.10)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
